Hash tag function improved. Link function added
I have added a website link javascript.  This script will search for
URLs in messages and turn them into a link to a website using a similar
method to the hash tag function.  Whilst I was doing this I realised
that there was a slight bug in the hash tag function which I have now
fixed.